feat(har): introduce urlFilter
This is a glob or regex pattern that filters entries recorder in the HAR.
