+++ packages/playwright-core/src/client/clientHelper.ts
@@ -17,7 +17,7 @@
 
 import type * as types from './types';
 import fs from 'fs';
-import { isString, isRegExp, constructURLBasedOnBaseURL } from '../utils';
+import { isString } from '../utils';
 
 export function envObjectToArray(env: types.Env): { name: string, value: string }[] {
   const result: { name: string, value: string }[] = [];
@@ -47,91 +47,4 @@ export async function evaluationScript(fun: Function | string | { path?: string,
     return source;
   }
   throw new Error('Either path or content property must be present');
-}
-
-export function parsedURL(url: string): URL | null {
-  try {
-    return new URL(url);
-  } catch (e) {
-    return null;
-  }
-}
-
-export function urlMatches(baseURL: string | undefined, urlString: string, match: types.URLMatch | undefined): boolean {
-  if (match === undefined || match === '')
-    return true;
-  if (isString(match) && !match.startsWith('*'))
-    match = constructURLBasedOnBaseURL(baseURL, match);
-  if (isString(match))
-    match = globToRegex(match);
-  if (isRegExp(match))
-    return match.test(urlString);
-  if (typeof match === 'string' && match === urlString)
-    return true;
-  const url = parsedURL(urlString);
-  if (!url)
-    return false;
-  if (typeof match === 'string')
-    return url.pathname === match;
-  if (typeof match !== 'function')
-    throw new Error('url parameter should be string, RegExp or function');
-  return match(url);
-}
-
-const escapeGlobChars = new Set(['/', '$', '^', '+', '.', '(', ')', '=', '!', '|']);
-
-export function globToRegex(glob: string): RegExp {
-  const tokens = ['^'];
-  let inGroup;
-  for (let i = 0; i < glob.length; ++i) {
-    const c = glob[i];
-    if (escapeGlobChars.has(c)) {
-      tokens.push('\\' + c);
-      continue;
-    }
-    if (c === '*') {
-      const beforeDeep = glob[i - 1];
-      let starCount = 1;
-      while (glob[i + 1] === '*') {
-        starCount++;
-        i++;
-      }
-      const afterDeep = glob[i + 1];
-      const isDeep = starCount > 1 &&
-          (beforeDeep === '/' || beforeDeep === undefined) &&
-          (afterDeep === '/' || afterDeep === undefined);
-      if (isDeep) {
-        tokens.push('((?:[^/]*(?:\/|$))*)');
-        i++;
-      } else {
-        tokens.push('([^/]*)');
-      }
-      continue;
-    }
-
-    switch (c) {
-      case '?':
-        tokens.push('.');
-        break;
-      case '{':
-        inGroup = true;
-        tokens.push('(');
-        break;
-      case '}':
-        inGroup = false;
-        tokens.push(')');
-        break;
-      case ',':
-        if (inGroup) {
-          tokens.push('|');
-          break;
-        }
-        tokens.push('\\' + c);
-        break;
-      default:
-        tokens.push(c);
-    }
-  }
-  tokens.push('$');
-  return new RegExp(tokens.join(''));
 }
